Move log module to rdiffbackup.singletons
