+++ src/rdiffbackup/locations/_repo_shadow.py
@@ -35,7 +35,6 @@ from rdiff_backup import (
     C,
     hash,
     iterfile,
-    log,
     Rdiff,
     robust,
     rorpiter,
@@ -50,7 +49,7 @@ from rdiffbackup.locations import fs_abilities, increment, location
 from rdiffbackup.locations.map import filenames as map_filenames
 from rdiffbackup.locations.map import hardlinks as map_hardlinks
 from rdiffbackup.locations.map import longnames as map_longnames
-from rdiffbackup.singletons import consts, generics, specifics
+from rdiffbackup.singletons import consts, generics, log, specifics
 from rdiffbackup.utils import locking, simpleps
 
 # ### COPIED FROM BACKUP ####
