+++ src/rdiffbackup/locations/map/longnames.py
@@ -38,7 +38,7 @@ it later.
 """
 
 import errno
-from rdiff_backup import log
+from rdiffbackup.singletons import log
 from rdiffbackup.utils import safestr
 
 _long_name_dir = b"long_filename_data"
