+++ src/rdiffbackup/meta/acl_posix.py
@@ -33,11 +33,11 @@ try:
 except ImportError:
     pass
 
-from rdiff_backup import C, log, rorpiter
+from rdiff_backup import C, rorpiter
 from rdiffbackup import meta
 from rdiffbackup.utils import usrgrp
 from rdiffbackup.locations.map import owners as map_owners
-from rdiffbackup.singletons import generics, specifics
+from rdiffbackup.singletons import generics, log, specifics
 
 # When an ACL gets dropped, put name in dropped_acl_names.  This is
 # only used so that only the first dropped ACL for any given name
