+++ src/rdiffbackup/meta/acl_win.py
@@ -18,9 +18,9 @@
 # 02110-1301, USA
 
 import os
-from rdiff_backup import C, log, rorpiter
+from rdiff_backup import C, rorpiter
 from rdiffbackup import meta
-from rdiffbackup.singletons import generics, specifics
+from rdiffbackup.singletons import generics, log, specifics
 from rdiffbackup.utils import safestr
 
 try:
